+++ src/opendr/perception/object_tracking_2d/__init__.py
@@ -1,5 +1,8 @@
-from opendr.perception.object_tracking_2d.fair_mot.object_tracking_2d_fair_mot_learner import ObjectTracking2DFairMotLearner
-from opendr.perception.object_tracking_2d.deep_sort.object_tracking_2d_deep_sort_learner import ObjectTracking2DDeepSortLearner
+from opendr.perception.object_tracking_2d.fair_mot.object_tracking_2d_fair_mot_learner import \
+    ObjectTracking2DFairMotLearner
+from opendr.perception.object_tracking_2d.deep_sort.object_tracking_2d_deep_sort_learner import \
+    ObjectTracking2DDeepSortLearner
+from opendr.perception.object_tracking_2d.siamrpn.siamrpn_learner import SiamRPNLearner
 
 from opendr.perception.object_tracking_2d.datasets.mot_dataset import (
     MotDataset,
@@ -14,4 +17,5 @@ from opendr.perception.object_tracking_2d.datasets.market1501_dataset import (
 )
 
 __all__ = ['ObjectTracking2DFairMotLearner', 'ObjectTracking2DDeepSortLearner', 'MotDataset', 'MotDatasetIterator',
-           'RawMotDatasetIterator', 'RawMotWithDetectionsDatasetIterator', 'Market1501Dataset', 'Market1501DatasetIterator']
+           'RawMotDatasetIterator', 'RawMotWithDetectionsDatasetIterator', 'Market1501Dataset', 'Market1501DatasetIterator',
+           'SiamRPNLearner']